Implement Twig runtimes and use Twig namespaces
